Add note to README to update workflow triggers
This was an oversight on my part that I would suspect others might run into - I used the template and pushed to a main branch, and totally missed the fact that none of the workflows were running. When it was time to properly add tests I realized that the files were there, and I simply needed to update the branch from master to main. So that others don't run into this issue, I am suggesting adding a line in the README to alert the reader to do the change.